migrations, seeders and controller fixed

+++ database/migrations/2021_01_29_001224_category_product_table.php
@@ -14,8 +14,8 @@ class CategoryProductTable extends Migration
     public function up()
     {
         Schema::create('category_product', function (Blueprint $table) {
-            $table->integer('category_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->unsignedBigInteger('category_id')->unsigned();
+            $table->unsignedBigInteger('product_id')->unsigned();
 
             /**
              * indicando que son foraneas
